+++ src/main/java/com/bookzone/controller/HomeController.java
@@ -11,10 +11,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 
 import com.bookzone.service.RegistrationService;
 
@@ -38,7 +35,7 @@ public class LibrarianController {
 	 * @return name of the Index page
 	 */
 	@GetMapping("/")
-	public String goToIndex() {
+	public String showIndex() {
 		librarianControllerLogger.info("LibrarianControllerLogger: Currently at Index page");
 		return "index";
 	}
@@ -52,39 +49,6 @@ public class LibrarianController {
 	public String goToHome() {
 		librarianControllerLogger.info("LibrarianControllerLogger: Currently at Home page");
 		return "home";
-	}
-	
-	/**
-	 * Goes to the Login page
-	 * 
-	 * @return name of the Login page
-	 */
-	@GetMapping("/login")
-	public String goToLogin() {
-		librarianControllerLogger.info("LibrarianControllerLogger: Currently at Login page");
-		return "login";
-	}
-	
-	/**
-	 * Manages a Librarian's login process
-	 * 
-	 * @param email: email address of the Librarian
-	 * @param password: password of the Librarian
-	 * @param model: The model where attributes can be added for the view
-	 * @return Redirection to the Home page for a successful login, or back to the login page
-	 * with an error message displayed for unsuccessful login
-	 */
-	@PostMapping("/login")
-	public String loginLibrarian(@RequestParam String email, @RequestParam String password, Model model) {
-	    boolean doesLibrarianExist = this.registrationService.loginLibrarian(email, password);
-	    if (doesLibrarianExist) {
-	        librarianControllerLogger.info("LibrarianControllerLogger: Librarian successfully logged in, proceeding to Home page");
-	        return "redirect:/home";
-	    } else {
-	        librarianControllerLogger.error("LibrarianControllerLogger: Librarian unsuccessful in logging in, proceeding to Login page");
-	        model.addAttribute("error", "Invalid email or password. Please try again.");
-	        return "login";
-	    }
 	}
 
 	/**
