+++ src/main/java/com/bookzone/controller/IndexController.java
@@ -1,0 +1,2 @@
+package com.bookzone.controller;public class IndexController {
+}
