+++ src/main/java/com/bookzone/controller/LoginController.java
@@ -1,0 +1,2 @@
+package com.bookzone.controller;public class LoginController {
+}
