+++ src/main/java/com/bookzone/service/LoginService.java
@@ -1,0 +1,2 @@
+package com.bookzone.service;public class LoginService {
+}
