+++ src/test/java/com/bookzone/controller/ErrorPageControllerTest.java
@@ -1,4 +1,4 @@
-package com.bookzone.controllerTest;
+package com.bookzone.controller;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -7,8 +7,6 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-
-import com.bookzone.controller.ErrorPageController;
 
 @ExtendWith(MockitoExtension.class)
 class ErrorPageControllerTest {
