+++ src/test/java/com/bookzone/controller/IndexControllerTest.java
@@ -1,0 +1,19 @@
+package com.bookzone.controller;
+
+import org.junit.jupiter.api.Test;
+import org.mockito.Mock;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class IndexControllerTest {
+
+    @Mock
+    private IndexController indexController;
+
+    @Test
+    void test_showIndex() {
+        String result = indexController.showIndex();
+        assertEquals("index", result);
+    }
+
+}
