+++ src/test/java/com/bookzone/model/BookTest.java
@@ -1,4 +1,4 @@
-package com.bookzone.modelTest;
+package com.bookzone.model;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
@@ -7,108 +7,106 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
 
-import com.bookzone.model.SpecialBook;
-
-class SpecialBookTest {
+class BookTest {
 
 	@Mock
-	SpecialBook book1;
+	Book book1;
 	
 	@Mock
-	SpecialBook book2;
+	Book book2;
 	
 	@BeforeEach
 	void setUp() throws Exception {
-		book1 = new SpecialBook(1, "King: A Life", "R.F.Kuang", "Fiction", 2023);
-		book2 = new SpecialBook(2, "The Bad Beginning", "Daniel Handler", "Novel", 1999);
+		book1 = new Book(1, "King: A Life", "R.F. Kuang", "Fiction", 2023);
+		book2 = new Book(2, "The Bad Beginning", "Daniel Handler", "Novel", 1999);
 	}
 
 	@Test
-	void checksTitle_returnsCorrectTitle() {
+	void testTitle_returnsCorrectTitle() {
 		
 		assertEquals("King: A Life", book1.getTitle());
 		
 	}
 	
 	@Test
-	void checksTitle_doesNotReturnIncorrectTitle() {
+	void testTitle_doesNotReturnIncorrectTitle() {
 		
 		assertNotEquals(book2.getTitle(), book1.getTitle());
 		
 	}
 	
 	@Test
-	void checksTitle_doesNotReturnNullForCorrectTitle() {
+	void testTitle_doesNotReturnNullForCorrectTitle() {
 		
 		assertNotEquals(null, book1.getAuthor());
 		
 	}
 	
 	@Test
-	void checksAuthor_returnsCorrectAuthor() {
+	void testAuthor_returnsCorrectAuthor() {
 	
 		assertEquals("Daniel Handler", book2.getAuthor());
 		
 	}
 	
 	@Test
-	void checksAuthor_doesNotReturnIncorrectAuthor() {
+	void testAuthor_doesNotReturnIncorrectAuthor() {
 		
 		assertNotEquals(book1.getAuthor(), book2.getAuthor());
 		
 	}
 	
 	@Test
-	void checksAuthor_doesNotReturnNullForCorrectAuthor() {
+	void testAuthor_doesNotReturnNullForCorrectAuthor() {
 		
 		assertNotEquals(null, book1.getAuthor());
 		
 	}
 	
 	@Test
-	void checksYear_returnsCorrectYear() {
+	void testYear_returnsCorrectYear() {
 		
 		assertEquals(2023, book1.getYear());
 		
 	}
 	
 	@Test
-	void checksYear_doesNotReturnIncorrectPrice() {
+	void testYear_doesNotReturnIncorrectPrice() {
 		
 		assertNotEquals(book2.getYear(), book1.getYear());
 		
 	}
 	
 	@Test
-	void checksCategory_returnsCorrectCategory() {
+	void testCategory_returnsCorrectCategory() {
 		
 		assertEquals("Novel", book2.getCategory());
 		
 	}
 	
 	@Test
-	void checksCategory_doesNotReturnIncorrectCategory() {
+	void testCategory_doesNotReturnIncorrectCategory() {
 		
 		assertNotEquals(book1.getCategory(), book2.getCategory());
 		
 	}
 	
 	@Test
-	void checksCategory_doesNotReturnNullForCorrectUser() {
+	void testCategory_doesNotReturnNullForCorrectUser() {
 		
 		assertNotEquals(null, book2.getCategory());
 		
 	}
 	
 	@Test
-	void checksId_returnsCorrectId() {
+	void testId_returnsCorrectId() {
 		
 		assertEquals(1, book1.getId());
 		
 	}
 	
 	@Test
-	void checksId_doesNotReturnIncorrectId() {
+	void testId_doesNotReturnIncorrectId() {
 		
 		assertNotEquals(book2.getId(), book1.getId());
