+++ src/test/java/com/bookzone/model/SpecialBookTest.java
@@ -1,4 +1,4 @@
-package com.bookzone.modelTest;
+package com.bookzone.model;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
@@ -7,108 +7,106 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
 
-import com.bookzone.model.Book;
-
-class BookTest {
+class SpecialBookTest {
 
 	@Mock
-	Book book1;
+	SpecialBook book1;
 	
 	@Mock
-	Book book2;
+	SpecialBook book2;
 	
 	@BeforeEach
 	void setUp() throws Exception {
-		book1 = new Book(1, "King: A Life", "R.F. Kuang", "Fiction", 2023);
-		book2 = new Book(2, "The Bad Beginning", "Daniel Handler", "Novel", 1999);
+		book1 = new SpecialBook(1, "King: A Life", "R.F.Kuang", "Fiction", 2023);
+		book2 = new SpecialBook(2, "The Bad Beginning", "Daniel Handler", "Novel", 1999);
 	}
 
 	@Test
-	void testTitle_returnsCorrectTitle() {
+	void checksTitle_returnsCorrectTitle() {
 		
 		assertEquals("King: A Life", book1.getTitle());
 		
 	}
 	
 	@Test
-	void testTitle_doesNotReturnIncorrectTitle() {
+	void checksTitle_doesNotReturnIncorrectTitle() {
 		
 		assertNotEquals(book2.getTitle(), book1.getTitle());
 		
 	}
 	
 	@Test
-	void testTitle_doesNotReturnNullForCorrectTitle() {
+	void checksTitle_doesNotReturnNullForCorrectTitle() {
 		
 		assertNotEquals(null, book1.getAuthor());
 		
 	}
 	
 	@Test
-	void testAuthor_returnsCorrectAuthor() {
+	void checksAuthor_returnsCorrectAuthor() {
 	
 		assertEquals("Daniel Handler", book2.getAuthor());
 		
 	}
 	
 	@Test
-	void testAuthor_doesNotReturnIncorrectAuthor() {
+	void checksAuthor_doesNotReturnIncorrectAuthor() {
 		
 		assertNotEquals(book1.getAuthor(), book2.getAuthor());
 		
 	}
 	
 	@Test
-	void testAuthor_doesNotReturnNullForCorrectAuthor() {
+	void checksAuthor_doesNotReturnNullForCorrectAuthor() {
 		
 		assertNotEquals(null, book1.getAuthor());
 		
 	}
 	
 	@Test
-	void testYear_returnsCorrectYear() {
+	void checksYear_returnsCorrectYear() {
 		
 		assertEquals(2023, book1.getYear());
 		
 	}
 	
 	@Test
-	void testYear_doesNotReturnIncorrectPrice() {
+	void checksYear_doesNotReturnIncorrectPrice() {
 		
 		assertNotEquals(book2.getYear(), book1.getYear());
 		
 	}
 	
 	@Test
-	void testCategory_returnsCorrectCategory() {
+	void checksCategory_returnsCorrectCategory() {
 		
 		assertEquals("Novel", book2.getCategory());
 		
 	}
 	
 	@Test
-	void testCategory_doesNotReturnIncorrectCategory() {
+	void checksCategory_doesNotReturnIncorrectCategory() {
 		
 		assertNotEquals(book1.getCategory(), book2.getCategory());
 		
 	}
 	
 	@Test
-	void testCategory_doesNotReturnNullForCorrectUser() {
+	void checksCategory_doesNotReturnNullForCorrectUser() {
 		
 		assertNotEquals(null, book2.getCategory());
 		
 	}
 	
 	@Test
-	void testId_returnsCorrectId() {
+	void checksId_returnsCorrectId() {
 		
 		assertEquals(1, book1.getId());
 		
 	}
 	
 	@Test
-	void testId_doesNotReturnIncorrectId() {
+	void checksId_doesNotReturnIncorrectId() {
 		
 		assertNotEquals(book2.getId(), book1.getId());
