+++ src/test/java/com/bookzone/repository/PersonRepositoryTest.java
@@ -1,8 +1,7 @@
-package com.bookzone.repositoryTest;
+package com.bookzone.repository;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
 
 import java.util.Optional;
 
@@ -13,7 +12,6 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import com.bookzone.model.Librarian;
-import com.bookzone.repository.PersonRepository;
 
 @ExtendWith(MockitoExtension.class)
 class PersonRepositoryTest {
