+++ src/test/java/com/bookzone/service/BookServiceTest.java
@@ -1,4 +1,4 @@
-package com.bookzone.serviceTest;
+package com.bookzone.service;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.doNothing;
@@ -19,7 +19,6 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 import com.bookzone.model.Book;
 import com.bookzone.repository.BookRepository;
-import com.bookzone.service.BookService;
 
 @ExtendWith(MockitoExtension.class)
 class BookServiceTest {
